Extract Interface from PodcastService
 - The impl-class is stored in the Scanner package
 - Since this is a variant of scanning, the implementation needs to
   understand how scanning works.
 - Some deatures are scans and exclusive relationships. Scan's locking
   mechanism is encapsulated in the scanner package
